tests: add tests for [index] and .key syntaxes
We added [index] support for consistency, however due to my personal
laziness .key is accepted as a side effect of this change.
Most implementations, except the implementation used by kubectl, do not
support it, however I think it will not introduce any unexpected
drawback.
I added it to the tests since we want to keep this library as stable as
possible, even for unexpected features.